+++ software_esp32/include/compile_time.h
@@ -79,4 +79,3 @@
 #endif /* COMPILE_TIME_H_ */
 
 
-