+++ software_esp32/include/stmApp.h
@@ -57,7 +57,8 @@ typedef struct  {
   uint32_t  movements;
   uint32_t  opening_count;
   uint32_t  closing_count;
-  int32_t  deadzone_count;
+  int32_t   deadzone_count;
+  uint8_t   calibRetries;
 } ACTUATOR_STRUC;
 
 typedef struct  {
@@ -228,7 +229,7 @@ private:
   char*	  cmdptrend;
 
   #define noOfLargeArgs 2
-  #define noOfSmallArgs 8
+  #define noOfSmallArgs 10
   #define noOfArgs noOfLargeArgs+noOfSmallArgs
   #define cmdSize 50
   #define largeArgSize 1200
@@ -244,10 +245,13 @@ private:
         arg6[smallArgSize],
         arg7[smallArgSize],
         arg8[smallArgSize],
-        arg9[smallArgSize];
+        arg9[smallArgSize],
+        arg10[smallArgSize],
+        arg11[smallArgSize];
 
-  char *argptr[noOfArgs]={arg0,arg1,arg2,arg3,arg4,arg5,arg6,arg7,arg8,arg9};
-  uint16_t argSize[noOfArgs]={sizeof(arg0),sizeof(arg1),sizeof(arg2),sizeof(arg3),sizeof(arg4),sizeof(arg5),sizeof(arg6),sizeof(arg7),sizeof(arg8),sizeof(arg9)};
+
+  char *argptr[noOfArgs]={arg0,arg1,arg2,arg3,arg4,arg5,arg6,arg7,arg8,arg9,arg10,arg11};
+  uint16_t argSize[noOfArgs]={sizeof(arg0),sizeof(arg1),sizeof(arg2),sizeof(arg3),sizeof(arg4),sizeof(arg5),sizeof(arg6),sizeof(arg7),sizeof(arg8),sizeof(arg9),sizeof(arg10),sizeof(arg11)};
 
 	uint8_t	argcnt;