+++ software_esp32/src/mqtt.cpp
@@ -663,9 +663,10 @@ void CMqtt::publish_valves () {
 
                     if (StmApp.actuators[x].calibration) {
                         topicstr[len] = '\0';
-                        strncat(topicstr, "/calibration",sizeof(topicstr) - strlen (topicstr) - 1);
+                        strncat(topicstr, "/calibration/date",sizeof(topicstr) - strlen (topicstr) - 1);
                         publishValue(topicstr, (char*) VdmSystem.localTime().c_str());
                     }
+
                     if ((VdmConfig.configFlash.valvesControlConfig.valveControlConfig[x].link>0) && (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[x].link<13)) {
                         topicstr[len] = '\0';
                         strncat(topicstr, "/control/link",sizeof(topicstr) - strlen (topicstr) - 1);
@@ -674,14 +675,57 @@ void CMqtt::publish_valves () {
                     }
                     lastValveValues[x].state=valveStates[x].thisState;
                 }
-                // meancurrent
-                if ((lastValveValues[x].meanCurrent!=StmApp.actuators[x].meancurrent) || forcePublish || lastValveValues[x].publishTimeOut) {
+
+                // calibration retries
+                if ((lastValveValues[x].calibRetries!=StmApp.actuators[x].calibRetries) || forcePublish || lastValveValues[x].publishTimeOut) {
                     topicstr[len] = '\0';
-                    strncat(topicstr, "/meancur",sizeof(topicstr) - strlen (topicstr) - 1);
-                    itoa(StmApp.actuators[x].meancurrent, valstr, 10);
+                    strncat(topicstr, "/calibration/retries",sizeof(topicstr) - strlen (topicstr) - 1);
+                    itoa(StmApp.actuators[x].calibRetries, valstr, 10);
                     publishValue(topicstr, valstr);
-                    lastValveValues[x].meanCurrent=StmApp.actuators[x].meancurrent;
+                    lastValveValues[x].calibRetries=StmApp.actuators[x].calibRetries;
                 }
+                // diag
+                if (VdmConfig.configFlash.protConfig.protocolFlags.publishDiag) {
+                    // meancurrent
+                    if ((lastValveValues[x].meanCurrent!=StmApp.actuators[x].meancurrent) || forcePublish || lastValveValues[x].publishTimeOut) {
+                        topicstr[len] = '\0';
+                        strncat(topicstr, "/diag/meanCurrrent",sizeof(topicstr) - strlen (topicstr) - 1);
+                        itoa(StmApp.actuators[x].meancurrent, valstr, 10);
+                        publishValue(topicstr, valstr);
+                        lastValveValues[x].meanCurrent=StmApp.actuators[x].meancurrent;
+                    }
+                    // counts
+                    if ((lastValveValues[x].openCount!=StmApp.actuators[x].opening_count) || forcePublish || lastValveValues[x].publishTimeOut) {
+                        topicstr[len] = '\0';
+                        strncat(topicstr, "/diag/openCounts",sizeof(topicstr) - strlen (topicstr) - 1);
+                        itoa(StmApp.actuators[x].opening_count, valstr, 10);
+                        publishValue(topicstr, valstr);
+                        lastValveValues[x].openCount=StmApp.actuators[x].opening_count;
+                    }
+                    if ((lastValveValues[x].closeCount!=StmApp.actuators[x].closing_count) || forcePublish || lastValveValues[x].publishTimeOut) {
+                        topicstr[len] = '\0';
+                        strncat(topicstr, "/diag/closeCounts",sizeof(topicstr) - strlen (topicstr) - 1);
+                        itoa(StmApp.actuators[x].closing_count, valstr, 10);
+                        publishValue(topicstr, valstr);
+                        lastValveValues[x].closeCount=StmApp.actuators[x].closing_count;
+                    }
+                    if ((lastValveValues[x].deadzoneCount!=StmApp.actuators[x].deadzone_count) || forcePublish || lastValveValues[x].publishTimeOut) {
+                        topicstr[len] = '\0';
+                        strncat(topicstr, "/diag/deadZoneCounts",sizeof(topicstr) - strlen (topicstr) - 1);
+                        itoa(StmApp.actuators[x].deadzone_count, valstr, 10);
+                        publishValue(topicstr, valstr);
+                        lastValveValues[x].deadzoneCount=StmApp.actuators[x].deadzone_count;
+                    }
+                    // moves
+                    if ((lastValveValues[x].movements!=StmApp.actuators[x].movements) || forcePublish || lastValveValues[x].publishTimeOut) {
+                        topicstr[len] = '\0';
+                        strncat(topicstr, "/diag/moves",sizeof(topicstr) - strlen (topicstr) - 1);
+                        itoa(StmApp.actuators[x].movements, valstr, 10);
+                        publishValue(topicstr, valstr);
+                        lastValveValues[x].movements=StmApp.actuators[x].movements;
+                    }
+                }
+
                 // temperature 1st sensor
                 if ((StmApp.stmInitState==STM_INIT_FINISHED) && StmApp.oneWireAllRead) {
                     if (StmApp.actuators[x].tIdx1>0) {