+++ software_esp32/src/stmApp.cpp
@@ -50,6 +50,7 @@
 #include "VdmNet.h"
 #include "VdmSystem.h"
 #include "VdmConfig.h"
+#include "VdmTask.h"
 #include "stm32.h"
 #include "Queue.h"
 #include "Messenger.h"
@@ -119,6 +120,7 @@ void  CStmApp::app_setup() {
         actuators[x].tIdx2 = 0;
         actuators[x].lastState = actuators[x].state;
         actuators[x].worked = false;
+        actuators[x].calibRetries = 0;
    }
 
     for (uint8_t x = 0; x<ACTUATOR_COUNT; x++) {
@@ -552,13 +554,15 @@ void  CStmApp::app_check_data()
                     actuators[idx].calibration = (val8>=0x80);
                     actuators[idx].temp1 = ConvertCF(atoi(argptr[4]))+getTOffset(actuators[idx].tIdx1);
                     actuators[idx].temp2 =  ConvertCF(atoi(argptr[5]))+getTOffset(actuators[idx].tIdx2);
-                    if (argcnt == 10) {
+                    if (argcnt >= 10) {
                         actuators[idx].movements = atoi(argptr[6]);   
                         actuators[idx].opening_count = atoi(argptr[7]);
                         actuators[idx].closing_count = atoi(argptr[8]);
                         actuators[idx].deadzone_count = atoi(argptr[9]); 
                     }
-
+                    if (argcnt >= 11) {
+                        actuators[idx].calibRetries = atoi(argptr[10]); 
+                    }
                     if (VdmConfig.configFlash.netConfig.syslogLevel>=VISMODE_DETAIL) {
                         syslog.log(LOG_DEBUG, "STMApp:got valve data #"+String(argptr[0])+" pos:"+String(argptr[1])+
                         " mean:"+String(argptr[2])+" state:"+String(argptr[3])+" t1:"+String(argptr[4])+" t2:"+String(argptr[5]));
@@ -639,7 +643,7 @@ void  CStmApp::app_check_data()
             }
             if (tempsPrivCount==0) {
                 fastGetOneWire=false;
-                oneWireAllRead=true;
+               // oneWireAllRead=true;
             }
             appState=APP_IDLE;
         }
@@ -663,7 +667,7 @@ void  CStmApp::app_check_data()
                 if (tempIndex>=tempsCount) {  // all temp sensors read
                     tempIndex=0;
                     fastGetOneWire=false;
-                    oneWireAllRead=true;
+                    if (VdmTask.piTaskInitiated) oneWireAllRead=true;
                 }
                 #ifdef AppDebug
                     UART_DBG.println("read onwire "+String(tempIndex)+":"+String(fastGetOneWire));
@@ -678,7 +682,7 @@ void  CStmApp::app_check_data()
                 if (tempIndex>=tempsCount) {  // all temp sensors read
                     tempIndex=0;
                     fastGetOneWire=false;
-                    oneWireAllRead=true;
+                    if (VdmTask.piTaskInitiated) oneWireAllRead=true;
                 }
             }
             appState=APP_IDLE;